requirements.txt
PyQt5
pillow
imagehash
python-pptx
pywin32
